+++ CPU Emulator SDL2/other/vec3.h
@@ -20,15 +20,15 @@ using std::fabs;
 class vec3 {
     public:
         vec3() : e{0,0,0} {}
-        vec3(double e0, double e1, double e2) : e{e0, e1, e2} {}
+        vec3(long double e0, long double e1, long double e2) : e{e0, e1, e2} {}
 
-        double x() const { return e[0]; }
-        double y() const { return e[1]; }
-        double z() const { return e[2]; }
+        long double x() const { return e[0]; }
+        long double y() const { return e[1]; }
+        long double z() const { return e[2]; }
 
         vec3 operator-() const { return vec3(-e[0], -e[1], -e[2]); }
-        double operator[](int i) const { return e[i]; }
-        double& operator[](int i) { return e[i]; }
+        long double operator[](int i) const { return e[i]; }
+        long double& operator[](int i) { return e[i]; }
 
         vec3& operator+=(const vec3 &v) {
             e[0] += v.e[0];
@@ -37,22 +37,22 @@ class vec3 {
             return *this;
         }
 
-        vec3& operator*=(const double t) {
+        vec3& operator*=(const long double t) {
             e[0] *= t;
             e[1] *= t;
             e[2] *= t;
             return *this;
         }
 
-        vec3& operator/=(const double t) {
+        vec3& operator/=(const long double t) {
             return *this *= 1/t;
         }
 
-        double length() const {
+        long double length() const {
             return sqrt(length_squared());
         }
 
-        double length_squared() const {
+        long double length_squared() const {
             return e[0]*e[0] + e[1]*e[1] + e[2]*e[2];
         }
 
@@ -64,7 +64,7 @@ class vec3 {
 
 
     public:
-        double e[3];
+        long double e[3];
 };
 
 
@@ -91,19 +91,19 @@ inline vec3 operator*(const vec3 &u, const vec3 &v) {
     return vec3(u.e[0] * v.e[0], u.e[1] * v.e[1], u.e[2] * v.e[2]);
 }
 
-inline vec3 operator*(double t, const vec3 &v) {
+inline vec3 operator*(long double t, const vec3 &v) {
     return vec3(t*v.e[0], t*v.e[1], t*v.e[2]);
 }
 
-inline vec3 operator*(const vec3 &v, double t) {
+inline vec3 operator*(const vec3 &v, long double t) {
     return t * v;
 }
 
-inline vec3 operator/(vec3 v, double t) {
+inline vec3 operator/(vec3 v, long double t) {
     return (1/t) * v;
 }
 
-inline double dot(const vec3 &u, const vec3 &v) {
+inline long double dot(const vec3 &u, const vec3 &v) {
     return u.e[0] * v.e[0]
          + u.e[1] * v.e[1]
          + u.e[2] * v.e[2];
@@ -123,7 +123,7 @@ inline vec3 reflect(const vec3& v, const vec3& n) {
     return v - 2*dot(v,n)*n;
 }
 
-inline vec3 refract(const vec3& uv, const vec3& n, double etai_over_etat) {
+inline vec3 refract(const vec3& uv, const vec3& n, long double etai_over_etat) {
     auto cos_theta = fmin(dot(-uv, n), 1.0);
     vec3 r_out_perp =  etai_over_etat * (uv + cos_theta*n);
     vec3 r_out_parallel = -sqrt(fabs(1.0 - r_out_perp.length_squared())) * n;
